Extract command config to module
